requirements.txt
asgiref==3.10.0
diff-match-patch==20241021
Django==5.2.8
django-asset-definitions==1.0.0
django-cors-headers==4.9.0
django-dashboards==0.1.7
django-import-export==4.3.14
narwhals==2.12.0
numpy==2.3.5
packaging==25.0
pandas==2.3.3
patsy==1.0.2
pillow==12.0.0
plotly==6.4.0
plotly-express==0.4.1
python-dateutil==2.9.0.post0
pytz==2025.2
scipy==1.16.3
six==1.17.0
sqlparse==0.5.3
statsmodels==0.14.5
tablib==3.9.0
typing_extensions==4.15.0
tzdata==2025.2
